Cleanup, add kilocode_markers, remove reason_details
