+++ src/api/transform/stream.ts
@@ -1,15 +1,18 @@
 import { ApiStreamNativeToolCallsChunk } from "./kilocode/api-stream-native-tool-calls-chunk"
+import { ReasoningDetail } from "./kilocode/reasoning-details"
 
 export type ApiStream = AsyncGenerator<ApiStreamChunk>
 
 export type ApiStreamChunk =
+	// kilocode_change start
+	| ApiStreamNativeToolCallsChunk
+	| ApiStreamReasoningDetailsChunk
+	// kilocode_change end
 	| ApiStreamTextChunk
 	| ApiStreamUsageChunk
-	| ApiStreamNativeToolCallsChunk // kilocode_change
 	| ApiStreamReasoningChunk
 	| ApiStreamGroundingChunk
 	| ApiStreamError
-	| ApiStreamReasoningDetailsChunk
 
 export interface ApiStreamError {
 	type: "error"
@@ -29,8 +32,9 @@ export interface ApiStreamReasoningChunk {
 
 export interface ApiStreamReasoningDetailsChunk {
 	type: "reasoning_details"
-	reasoning_details: any // openrouter has various properties that we can pass back unmodified in api requests to preserve reasoning traces
+	reasoning_details: ReasoningDetail
 }
+// kilocode_change start
 
 export interface ApiStreamUsageChunk {
 	type: "usage"
