+++ src/core/condense/index.ts
@@ -6,6 +6,7 @@ import { t } from "../../i18n"
 import { ApiHandler } from "../../api"
 import { ApiMessage } from "../task-persistence/apiMessages"
 import { maybeRemoveImageBlocks } from "../../api/transform/image-cleaning"
+import { maybeRemoveReasoningDetails_kilocode } from "../../api/transform/kilocode/reasoning-details"
 
 export const N_MESSAGES_TO_KEEP = 3
 export const MIN_CONDENSE_THRESHOLD = 5 // Minimum percentage of context window to trigger condensing
@@ -134,8 +135,12 @@ export async function summarizeConversation(
 		content: "Summarize the conversation so far, as described in the prompt instructions.",
 	}
 
-	const requestMessages = maybeRemoveImageBlocks([...messagesToSummarize, finalRequestMessage], apiHandler).map(
-		({ role, content }) => ({ role, content }),
+	const requestMessages = maybeRemoveReasoningDetails_kilocode(
+		maybeRemoveImageBlocks([...messagesToSummarize, finalRequestMessage], apiHandler).map(({ role, content }) => ({
+			role,
+			content,
+		})),
+		undefined,
 	)
 
 	// Note: this doesn't need to be a stream, consider using something like apiHandler.completePrompt
