+++ src/core/task/Task.ts
@@ -124,6 +124,7 @@ import { MessageQueueService } from "../message-queue/MessageQueueService"
 import { AutoApprovalHandler } from "./AutoApprovalHandler"
 import { isAnyRecognizedKiloCodeError, isPaymentRequiredError } from "../../shared/kilocode/errorUtils"
 import { getAppUrl } from "@roo-code/types"
+import { maybeRemoveReasoningDetails_kilocode, ReasoningDetail } from "../../api/transform/kilocode/reasoning-details"
 
 const MAX_EXPONENTIAL_BACKOFF_SECONDS = 600 // 10 minutes
 const DEFAULT_USAGE_COLLECTION_TIMEOUT_MS = 5000 // 5 seconds
@@ -2062,11 +2063,14 @@ export class Task extends EventEmitter<TaskEvents> implements TaskLike {
 				// limit error, which gets thrown on the first chunk).
 				const stream = this.attemptApiRequest()
 				let assistantMessage = ""
-				let assistantToolUses = new Array<Anthropic.Messages.ToolUseBlockParam>() // kilocode_change
 				let reasoningMessage = ""
-				const reasoningDetails = []
 				let pendingGroundingSources: GroundingSource[] = []
 				this.isStreaming = true
+
+				// kilocode_change start
+				const assistantToolUses = new Array<Anthropic.Messages.ToolUseBlockParam>()
+				const reasoningDetails = new Array<ReasoningDetail>()
+				// kilocode_change end
 
 				try {
 					const iterator = stream[Symbol.asyncIterator]()
@@ -2098,14 +2102,6 @@ export class Task extends EventEmitter<TaskEvents> implements TaskLike {
 								break
 							}
 							// for openrouter providers
-							case "reasoning_details":
-								// reasoning_details may be an array of 0 or 1 items depending on how openrouter returns it
-								if (Array.isArray(chunk.reasoning_details)) {
-									reasoningDetails.push(...chunk.reasoning_details)
-								} else {
-									reasoningDetails.push(chunk.reasoning_details)
-								}
-								break
 							case "usage":
 								inputTokens += chunk.inputTokens
 								outputTokens += chunk.outputTokens
@@ -2121,7 +2117,15 @@ export class Task extends EventEmitter<TaskEvents> implements TaskLike {
 									pendingGroundingSources.push(...chunk.sources)
 								}
 								break
-							//kilocode_change start
+							// kilocode_change start
+							case "reasoning_details":
+								// reasoning_details may be an array of 0 or 1 items depending on how openrouter returns it
+								if (Array.isArray(chunk.reasoning_details)) {
+									reasoningDetails.push(...chunk.reasoning_details)
+								} else {
+									reasoningDetails.push(chunk.reasoning_details)
+								}
+								break
 							case "native_tool_calls": {
 								// Handle native OpenAI-format tool calls
 								// Process native tool calls through the parser
@@ -2144,7 +2148,7 @@ export class Task extends EventEmitter<TaskEvents> implements TaskLike {
 								presentAssistantMessage(this)
 								break
 							}
-							//kilocode_change end
+							// kilocode_change end
 							case "text": {
 								assistantMessage += chunk.text
 
@@ -2491,14 +2495,13 @@ export class Task extends EventEmitter<TaskEvents> implements TaskLike {
 						})
 					}
 
-					// kilocode_change start: also add tool calls to history with reasoning_details
+					// kilocode_change start: also add tool calls, reasoning_details to history
 					const assistantMessageContent = new Array<Anthropic.Messages.ContentBlockParam>()
-					if (assistantMessage) {
+					if (assistantMessage || reasoningDetails.length > 0) {
 						assistantMessageContent.push({
 							type: "text",
 							text: assistantMessage,
-							// reasoning_details only exists for openrouter providers
-							// @ts-ignore-next-line (reasoning_details is not a valid property for TextBlockParam)
+							// @ts-ignore-next-line OpenRouter-specific property
 							reasoning_details: reasoningDetails.length > 0 ? reasoningDetails : undefined,
 						})
 					}
@@ -2962,8 +2965,9 @@ export class Task extends EventEmitter<TaskEvents> implements TaskLike {
 		}
 
 		const messagesSinceLastSummary = getMessagesSinceLastSummary(this.apiConversationHistory)
-		let cleanConversationHistory = maybeRemoveImageBlocks(messagesSinceLastSummary, this.api).map(
-			({ role, content }) => ({ role, content }),
+		let cleanConversationHistory = maybeRemoveReasoningDetails_kilocode(
+			maybeRemoveImageBlocks(messagesSinceLastSummary, this.api).map(({ role, content }) => ({ role, content })),
+			apiConfiguration?.apiProvider,
 		)
 
 		// kilocode_change start
