--- src/utils/model-utils.ts
@@ -1,10 +1,0 @@
-/**
- * Determines if reasoning content should be skipped for a given model
- * Currently skips reasoning for Grok-4 models since they only display "thinking" without useful information
- */
-export function shouldSkipReasoningForModel(modelId?: string): boolean {
-	if (!modelId) {
-		return false
-	}
-	return modelId.includes("grok-4")
-}
